# internal/httpapi/server.go
package httpapi

import (
	"context"
	"io/fs"
	"log/slog"
	"net/http"
	"net/http/httputil"
	"net/url"
	"os"
	"strings"
	"sync"

	"github.com/danielgtaylor/huma/v2"
	"github.com/danielgtaylor/huma/v2/adapters/humachi"
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"

	"github.com/MateEke/picture-frame/internal/auth"
	"github.com/MateEke/picture-frame/internal/config"
	"github.com/MateEke/picture-frame/internal/files"
	"github.com/MateEke/picture-frame/internal/hostmetrics"
	"github.com/MateEke/picture-frame/internal/library"
	"github.com/MateEke/picture-frame/internal/power"
	"github.com/MateEke/picture-frame/internal/state"
	"github.com/MateEke/picture-frame/web"
)

// ScreenController is the manual screen-control surface used by the HTTP API,
// implemented by *display.Screen.
type ScreenController interface {
	On(ctx context.Context) error
	Off(ctx context.Context) error
	// Wake is On for a kiosk touch, but free on a lit panel: no write, no publish.
	Wake(ctx context.Context) error
	State() bool
	Auto() bool
	// Reconcile re-asserts desired power; the SSE handler calls it on connect
	// (a kiosk reconnect may follow a compositor restart).
	Reconcile(ctx context.Context)
}

// RotationController is implemented by display.Rotator; nil when the backend
// can't rotate.
type RotationController interface {
	Set(ctx context.Context, degrees int) error
	Reconcile(ctx context.Context)
	Supported() bool
}

// SlideshowController allows the HTTP layer to drive slideshow playback.
type SlideshowController interface {
	Next()
	Prev()
	RestartCycle()
}

// KioskBeater records a kiosk heartbeat (with the reporting build's version); implemented
// by *kioskwatch.Watch.
type KioskBeater interface {
	Beat(version string)
}

// SlidePlanner receives the kiosk's reported screen aspect (which rebuilds the
// slide plan); SetScreenAspect reports whether the value changed. Implemented by
// *slideplan.Planner.
type SlidePlanner interface {
	SetScreenAspect(aspect float64) bool
}

// SyncerStatus reports the current state of a remote library backend and lets
// the admin request an out-of-band sync. Aliased to the library-owned interface
// so producers (e.g. startup) can return it without importing this package.
type SyncerStatus = library.SyncerStatus

// Config holds the dependencies for the HTTP server.
type Config struct {
	Log         *slog.Logger
	Screen      ScreenController
	Rotator     RotationController // nil when the display backend can't rotate
	Bus         *state.Bus
	Library     *library.Library
	Slideshow   SlideshowController
	ImagesRoot  *os.Root
	KioskBeater KioskBeater // required
	// Aspect caches per-image dimensions; nil disables split-screen metadata.
	Aspect *library.AspectStore
	// Order persists the canonical image order; nil disables order saving.
	Order *library.OrderStore
	// Exclude persists which images are hidden from the slideshow; nil disables saving.
	Exclude *library.ExcludeStore
	// Thumbs serves gallery thumbnails; nil falls back to the full image.
	Thumbs *library.ThumbStore
	// Files stores non-image uploads; nil disables the /api/files routes (503).
	Files *files.Store
	// Planner rebuilds slide plans and receives the kiosk's screen aspect; nil in
	// tests/dev that don't exercise split-screen.
	Planner SlidePlanner
	// Backend is the active library backend, e.g. "fs" or "immich".
	Backend string
	// Syncer is non-nil for remote backends; exposes sync status to the UI.
	Syncer SyncerStatus
	// Updater is non-nil when the self-updater is wired; nil reads as no update available.
	Updater UpdaterStatus
	// WiFi is non-nil in production; nil in dev returns 503 for wifi routes.
	WiFi WiFiManager
	// Production serves the embedded build when true, else proxies to Vite.
	Production bool
	// WeatherActive gates the kiosk weather UI; see startup.WeatherEnabled.
	WeatherActive bool
	// BacklightSupported reports a /sys/class/backlight device (brightness control).
	BacklightSupported bool
	// SysfsBase roots device enumeration (Bluetooth adapters, display outputs);
	// "" defaults to /sys/class. Overridden in tests with a fake sysfs tree.
	SysfsBase string

	// Config editing
	Store         *config.Store     // persisted config state; nil in tests that don't exercise config
	RunningConfig config.Config     // initial snapshot of what live subsystems reflect (== saved at startup)
	LiveConfig    LiveConfig        // nil in dev/tests; applies Tier-1 changes in-process
	Restart       func() error      // nil in dev/tests; triggers an in-place re-exec.
	HostMetrics   HostMetricsReader // nil in tests; supplies Pi telemetry for the system card
	Power         PowerController   // nil in tests; reboots or powers off the host
}

// HostMetricsReader supplies Pi telemetry for the dashboard system card.
type HostMetricsReader interface {
	Read(ctx context.Context) hostmetrics.Metrics
}

// PowerController reboots or powers off the host. Capabilities reports what the
// polkit policy actually permits, so the UI can hide what would only fail.
type PowerController interface {
	Reboot(ctx context.Context) error
	PowerOff(ctx context.Context) error
	Capabilities() power.Capabilities
}

type server struct {
	log           *slog.Logger
	screen        ScreenController
	rotator       RotationController
	bus           *state.Bus
	lib           *library.Library
	slideshow     SlideshowController
	imagesRoot    *os.Root
	kioskBeater   KioskBeater
	aspect        *library.AspectStore
	order         *library.OrderStore
	exclude       *library.ExcludeStore
	thumbs        *library.ThumbStore
	files         *files.Store
	planner       SlidePlanner
	backend       string
	syncer        SyncerStatus
	updater       UpdaterStatus
	wifiMgr       WiFiManager
	sysfsBase     string
	weatherActive bool
	backlight     bool
	hostMetrics   HostMetricsReader
	power         PowerController
	auth          *auth.Authenticator
	authMu        sync.Mutex // serializes bcrypt work; see checkPasswordGated

	// cookie-less kiosk routes; see kioskExempt
	kioskPaths    map[string]bool
	kioskPrefixes []string

	store      *config.Store
	mu         sync.RWMutex // guards running only
	running    config.Config
	liveConfig LiveConfig
	restart    func() error
}

// NewServer constructs the root HTTP handler. With Production set it serves the
// embedded SvelteKit build (SPA fallback); otherwise it proxies to Vite.
func NewServer(cfg Config) http.Handler {
	backend := cfg.Backend
	if backend == "" {
		backend = config.BackendFS
	}
	sysfsBase := cfg.SysfsBase
	if sysfsBase == "" {
		sysfsBase = "/sys/class"
	}
	s := &server{
		log:           cfg.Log,
		screen:        cfg.Screen,
		rotator:       cfg.Rotator,
		bus:           cfg.Bus,
		lib:           cfg.Library,
		slideshow:     cfg.Slideshow,
		imagesRoot:    cfg.ImagesRoot,
		kioskBeater:   cfg.KioskBeater,
		aspect:        cfg.Aspect,
		order:         cfg.Order,
		exclude:       cfg.Exclude,
		thumbs:        cfg.Thumbs,
		files:         cfg.Files,
		planner:       cfg.Planner,
		backend:       backend,
		syncer:        cfg.Syncer,
		updater:       cfg.Updater,
		wifiMgr:       cfg.WiFi,
		sysfsBase:     sysfsBase,
		weatherActive: cfg.WeatherActive,
		backlight:     cfg.BacklightSupported,
		hostMetrics:   cfg.HostMetrics,
		power:         cfg.Power,
		auth:          auth.New(),
		store:         cfg.Store,
		running:       cfg.RunningConfig,
		liveConfig:    cfg.LiveConfig,
		restart:       cfg.Restart,
	}

	r := chi.NewRouter()
	r.Use(middleware.Recoverer)
	r.Use(s.requireAuth)

	humaConfig := huma.DefaultConfig("Picture Frame", "1.0.0")
	humaConfig.Info.Description = "Picture frame kiosk API"
	api := humachi.New(r, humaConfig)

	s.registerRoutes(api)
	// Streaming upload bypasses huma; see handleUploadFile.
	r.Post("/api/files", s.handleUploadFile)

	// OS captive-portal probe URLs, redirect to /admin/network when AP is active.
	for _, path := range captiveProbes {
		r.Get(path, s.handleCaptiveProbe)
	}

	// Legacy path; 302 not 301 so a stale bookmark can't outlive a future move.
	redirectLegacyWiFi := func(w http.ResponseWriter, req *http.Request) {
		http.Redirect(w, req, "/admin/network", http.StatusFound)
	}
	r.Get("/admin/wifi", redirectLegacyWiFi)
	r.Head("/admin/wifi", redirectLegacyWiFi) // chi Get doesn't match HEAD; it'd 200 off the SPA

	if cfg.Production {
		cfg.Log.Info("serving embedded build (production mode)")
		r.Handle("/*", newSPAHandler(cfg.Log))
	} else {
		cfg.Log.Info("proxying to Vite dev server (development mode)")
		r.Handle("/*", newViteProxy(cfg.Log))
	}

	return r
}

func (s *server) registerRoutes(api huma.API) {
	s.registerScreenRoutes(api)
	s.registerLibraryRoutes(api)
	s.registerImageRoutes(api)
	s.registerFileRoutes(api)
	s.registerTouchRoutes(api)
	s.registerSlideshowRoutes(api)
	s.registerHeartbeatRoutes(api)
	s.registerWiFiRoutes(api)
	s.registerHealthRoutes(api)
	s.registerSSERoutes(api)
	s.registerConfigRoutes(api)
	s.registerDeviceRoutes(api)
	s.registerSystemInfoRoutes(api)
	s.registerPowerRoutes(api)
	s.registerUpdateRoutes(api)
	s.registerAuthRoutes(api)
}

func newSPAHandler(log *slog.Logger) http.Handler {
	sub, err := fs.Sub(web.BuildFS, "build")
	if err != nil {
		log.Error("failed to create sub FS from embedded build", "err", err)
		return http.NotFoundHandler()
	}
	fileServer := http.FileServer(http.FS(sub))
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		path := strings.TrimPrefix(r.URL.Path, "/")
		if path == "" {
			path = "index.html"
		}
		if _, err := fs.Stat(sub, path); err != nil {
			// The fallback serves index.html, never cacheable as the asset path.
			w.Header().Set("Cache-Control", "no-cache")
			r2 := r.Clone(r.Context())
			r2.URL.Path = "/"
			fileServer.ServeHTTP(w, r2)
			return
		}
		// embed.FS files have a zero modtime (no implicit validators), so set
		// caching explicitly: content-hashed assets forever, the rest revalidates.
		if strings.HasPrefix(path, "_app/immutable/") {
			w.Header().Set("Cache-Control", "public, max-age=31536000, immutable")
		} else {
			w.Header().Set("Cache-Control", "no-cache")
		}
		fileServer.ServeHTTP(w, r)
	})
}

func newViteProxy(log *slog.Logger) http.Handler {
	target, _ := url.Parse("http://localhost:5173")
	proxy := httputil.NewSingleHostReverseProxy(target)
	proxy.ErrorHandler = func(w http.ResponseWriter, r *http.Request, err error) {
		log.Error("vite proxy error", "path", r.URL.Path, "err", err)
		http.Error(w, "Vite dev server unavailable, run: cd web && npm run dev", http.StatusBadGateway)
	}
	return proxy
}