# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
      "FileItem": {
        "additionalProperties": false,
        "properties": {
          "added": {
            "description": "Upload time",
            "format": "date-time",
            "type": "string"
          },
          "mime": {
            "description": "Media type",
            "type": "string"
          },
          "name": {
            "description": "Stored filename",
            "type": "string"
          },
          "original": {
            "description": "Filename as uploaded",
            "type": "string"
          },
          "size": {
            "description": "Size in bytes",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "original",
          "size",
          "mime",
          "added"
        ],
        "type": "object"
      },
...
      "ListFilesOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/schemas/ListFilesOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileItem"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "free_bytes": {
            "description": "Space left on the storage, when known",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "files"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Picture frame kiosk API",
    "title": "Picture Frame",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/files": {
      "get": {
        "operationId": "list-files",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListFilesOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "List uploaded non-image files"
      }
    },
...
    "/api/heartbeat": {
      "post": {
        "operationId": "heartbeat",
        "parameters": [
          {
            "explode": false,
            "in": "query",
            "name": "version",
            "schema": {
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "aspect",
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Record kiosk heartbeat"
      }
    },
...
  }
}